Refractor Rule Group struct and fieldname
- Unify the same table "rule group" & "group"
- Fix old group out of date schema
- Update Group & GroupQuery to clean up code.

+++ module/qic-api/cu/service.go
@@ -185,7 +185,7 @@ func GetFlowGroup(enterprise string) ([]model.Group, error) {
 	if enterprise == "" {
 		return nil, nil
 	}
-	queryCondition := model.GroupQuery{EnterpriseID: &enterprise, Type: []int{Flow}}
+	queryCondition := model.GroupQuery{EnterpriseID: &enterprise, Type: []int8{Flow}}
 	return groupDao.Group(nil, queryCondition)
 }
